// ReactWeaver.Client/src/lib/api/schema.ts
export interface components {
  schemas: {
    CreateWeatherForecastRequest: {
      /** Format: date */
      date: string;
      /** Format: int32 */
      temperatureC: number | string;
      summary?: null | string;
    };
    ForgotPasswordRequest: {
      email: string;
    };
    LoginRequest: {
      email: string;
      password: string;
      twoFactorCode?: null | string;
      twoFactorRecoveryCode?: null | string;
    };
    ProblemDetails: {
      type?: null | string;
      title?: null | string;
      /** Format: int32 */
      status?: null | number | string;
      detail?: null | string;
      instance?: null | string;
    };
    RegisterRequest: {
      email: string;
      password: string;
    };
    ResendConfirmationEmailRequest: {
      email: string;
    };
    ResetPasswordRequest: {
      email: string;
      resetCode: string;
      newPassword: string;
    };
    UpdateUserInfoRequest: {
      newEmail: string;
      newPassword: string;
      oldPassword: string;
    };
    UpdateWeatherForecastRequest: {
      /** Format: date */
      date: string;
      /** Format: int32 */
      temperatureC: number | string;
      summary?: null | string;
    };
    UserResponse: {
      id: string;
      email: string;
      roles: string[];
      isEmailConfirmed: boolean;
    };
    WeatherForecastResponse: {
      id: string;
      /** Format: date */
      date: string;
      /** Format: int32 */
      temperatureC: number | string;
      summary: null | string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}